% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_dashboard_pins.R
\docType{data}
\name{dashboard_pins}
\alias{dashboard_pins}
\title{dashboard_pins}
\format{
A data frame with 16 rows and 3 variables:
\describe{
\item{ pin_order }{  integer }
\item{ indicator_id }{  character }
\item{ hypothesis_id }{  character }
}
}
\source{
Source
}
\usage{
dashboard_pins
}
\description{
Description.
}
\keyword{datasets}
