% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_dashboard_timeline.R
\docType{data}
\name{dashboard_timeline}
\alias{dashboard_timeline}
\title{dashboard_timeline}
\format{
A data frame with 103 rows and 26 variables:
\describe{
\item{ pin_order }{  integer }
\item{ hypothesis_id }{  character }
\item{ hypothesis_label }{  character }
\item{ evidence_role }{  character }
\item{ indicator_id }{  character }
\item{ indicator }{  character }
\item{ indicator_label }{  character }
\item{ evidence_label }{  character }
\item{ domain_group }{  character }
\item{ survey_year }{  numeric }
\item{ baseline_year }{  numeric }
\item{ latest_year }{  numeric }
\item{ value }{  numeric }
\item{ precision }{  numeric }
\item{ baseline_value }{  numeric }
\item{ change_from_baseline }{  numeric }
\item{ pct_change_from_baseline }{  numeric }
\item{ indexed_value }{  numeric }
\item{ progress_score }{  numeric }
\item{ progress_pct_from_baseline }{  numeric }
\item{ progress_status }{  character }
\item{ desirable_direction }{  numeric }
\item{ desired_direction_label }{  character }
\item{ value_unit }{  character }
\item{ value_axis_label }{  character }
\item{ tooltip }{  character }
}
}
\source{
Source
}
\usage{
dashboard_timeline
}
\description{
Description.
}
\keyword{datasets}
