% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_domain_summary.R
\docType{data}
\name{domain_summary}
\alias{domain_summary}
\title{domain_summary}
\format{
A data frame with 10 rows and 7 variables:
\describe{
\item{ domain_group }{  character }
\item{ resources }{  integer }
\item{ indicators }{  integer }
\item{ rows }{  integer }
\item{ min_year }{  numeric }
\item{ max_year }{  numeric }
\item{ latest_year }{  numeric }
}
}
\source{
Source
}
\usage{
domain_summary
}
\description{
Description.
}
\keyword{datasets}
