% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_hypothesis_map.R
\docType{data}
\name{hypothesis_map}
\alias{hypothesis_map}
\title{hypothesis_map}
\format{
A data frame with 23 rows and 6 variables:
\describe{
\item{ hypothesis_id }{  character }
\item{ hypothesis_label }{  character }
\item{ hypothesis_claim }{  character }
\item{ interpretation_rule }{  character }
\item{ evidence_role }{  character }
\item{ indicator_id }{  character }
}
}
\source{
Source
}
\usage{
hypothesis_map
}
\description{
Description.
}
\keyword{datasets}
