% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_indicator_catalogue.R
\docType{data}
\name{indicator_catalogue}
\alias{indicator_catalogue}
\title{indicator_catalogue}
\format{
A data frame with 1196 rows and 24 variables:
\describe{
\item{ indicator_id }{  character }
\item{ resource_name }{  character }
\item{ resource_position }{  character }
\item{ resource_id }{  character }
\item{ resource_file }{  character }
\item{ domain_group }{  character }
\item{ analysis_theme }{  character }
\item{ series_family }{  character }
\item{ sex_group }{  character }
\item{ county_available }{  logical }
\item{ indicator }{  character }
\item{ indicator_label }{  character }
\item{ indicator_type }{  character }
\item{ value_unit }{  character }
\item{ value_axis_label }{  character }
\item{ rows }{  integer }
\item{ year_count }{  integer }
\item{ years_available }{  character }
\item{ min_year }{  numeric }
\item{ latest_year }{  numeric }
\item{ breakdowns_available }{  character }
\item{ characteristics_available }{  character }
\item{ has_ci }{  logical }
\item{ has_denominator }{  logical }
}
}
\source{
Source
}
\usage{
indicator_catalogue
}
\description{
Description.
}
\keyword{datasets}
