% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_indicator_latest.R
\docType{data}
\name{indicator_latest}
\alias{indicator_latest}
\title{indicator_latest}
\format{
A data frame with 1195 rows and 14 variables:
\describe{
\item{ indicator_id }{  character }
\item{ indicator }{  character }
\item{ indicator_label }{  character }
\item{ domain_group }{  character }
\item{ resource_name }{  character }
\item{ survey_year }{  numeric }
\item{ value }{  numeric }
\item{ precision }{  numeric }
\item{ value_unit }{  character }
\item{ value_axis_label }{  character }
\item{ ci_low }{  numeric }
\item{ ci_high }{  numeric }
\item{ denominator_weighted }{  numeric }
\item{ denominator_unweighted }{  numeric }
}
}
\source{
Source
}
\usage{
indicator_latest
}
\description{
Description.
}
\keyword{datasets}
