% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_indicator_trends.R
\docType{data}
\name{indicator_trends}
\alias{indicator_trends}
\title{indicator_trends}
\format{
A data frame with 1195 rows and 15 variables:
\describe{
\item{ indicator_id }{  character }
\item{ indicator }{  character }
\item{ indicator_label }{  character }
\item{ domain_group }{  character }
\item{ resource_name }{  character }
\item{ precision }{  numeric }
\item{ baseline_year }{  numeric }
\item{ baseline_value }{  numeric }
\item{ latest_year }{  numeric }
\item{ latest_value }{  numeric }
\item{ change_from_baseline }{  numeric }
\item{ observations }{  integer }
\item{ value_unit }{  character }
\item{ value_axis_label }{  character }
\item{ pct_change_from_baseline }{  numeric }
}
}
\source{
Source
}
\usage{
indicator_trends
}
\description{
Description.
}
\keyword{datasets}
